BitSetter: Record the travel distance not the final position
This makes it possible to reset the Z zero position without having
to remember to rerun the Initial Tool macro.

When probing, compute the distance that the spindle had to travel
to trigger the probe. For a New Tool, offset the current Z zero
position by the difference between the last and the current
distances travelled.