WIP move to TLO for BitSetter
The big advantage that this brings is that you can use multiple
workspace coordinate systems which allows you to cut multiples of
a part at once.

The secondary advantage that this brings is that it allows you to
use a probe that you cannot measure on the BitSetter by having a
fixed TLO for the probe and synchronizing the BitSetter TLO with
that value (see the long comment at the end of the macro).