cncjs: Mitigate EEPROM writes
When a GRBL command which writes to the EEPROM actually does the
write, it has to disable all interrupts on the arduino board.  This
means that all serial data transmitted during the write will be lost!!

Try to avoid problems by:

  * Remove the use of soft limits, this adds a ton of EEPROM writes
    and requires that they be run in the middle of the macro.

  * Move G10 commands to the last line of each macro that uses them.

This means that WE aren't sending any serial data after an EEPROM change.

This doesn't make it 100% safe though as CNCJS itself is periodically
sending $G and ? commands which could be sent during this time which
could theoretically cause the *NEXT* command to be a cryptic error.