config: Disable soft limits
Soft limits are a pain to deal with in thep probing macros and
my attempts to disable/enable them on the fly discovered that
any command that causes an EEPROM write can break the serial
communication channel with CNCjs.

The most important check is really the Z axis because that is one
axis I can accidentally trigger when runing a job that will
definitely cause problems.